refactor : separate AlarmDropdown Component
